Add screen for edit profile
