+++ app/src/main/AndroidManifest.xml
@@ -15,6 +15,9 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="m">
         <activity
+            android:name=".view.profile.editprofile.EditActivity"
+            android:exported="false" />
+        <activity
             android:name=".view.main.MainActivity"
             android:exported="false"
             android:label="@string/title_activity_login"
