+++ app/src/main/java/com/brazhnik/fobres/view/profile/editprofile/EditActivity.kt
@@ -1,0 +1,11 @@
+package com.brazhnik.fobres.view.profile.editprofile
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class EditActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_edit)
+    }
+}
