+++ app/src/main/java/com/brazhnik/fobres/view/profile/editprofile/EditPresenter.kt
@@ -1,0 +1,4 @@
+package com.brazhnik.fobres.view.profile.editprofile
+
+class EditPresenter {
+}
