+++ app/src/main/java/com/brazhnik/fobres/view/profile/editprofile/EditView.kt
@@ -1,0 +1,4 @@
+package com.brazhnik.fobres.view.profile.editprofile
+
+interface EditView {
+}
